restructed cpp and java files
