+++ app/src/main/cpp/filter.cpp
