+++ app/src/main/cpp/kiwixsearcher.cpp
