+++ app/src/main/cpp/library.cpp
