+++ app/src/main/cpp/manager.cpp
