+++ app/src/main/java/org/kiwix/java_libkiwix/JNIKiwix.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (C) 2017 Matthieu Gautier <mgautier@kymeria.org>
+ * Copyright (C) 2013 Emmanuel Engelhart <kelson@kiwix.org>
+ * Copyright (C) 2017 Matthieu Gautier <mgautier@kymeria.fr>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU  General Public License as published by
@@ -17,11 +18,19 @@
  * MA 02110-1301, USA.
  */
 
-package org.kiwix.kiwixlib;
+package org.kiwix.java_libkiwix;
 
-public class JNIKiwixException extends Exception
+import android.content.Context;
+import com.getkeepsafe.relinker.ReLinker;
+import org.kiwix.java_libkiwix.JNIICU;
+
+public class JNIKiwix
 {
-  public JNIKiwixException(String message) {
-    super(message);
+  public JNIKiwix(final Context context){
+    ReLinker.loadLibrary(context, "kiwix");
+  }
+
+  public void setDataDirectory(String icuDataDir) {
+    JNIICU.setDataDirectory(icuDataDir);
   }
 }
