+++ app/src/main/java/org/kiwix/java_libkiwix/JNIKiwixString.java
@@ -17,9 +17,21 @@
  * MA 02110-1301, USA.
  */
 
-package org.kiwix.kiwixlib;
+package org.kiwix.java_libkiwix;
 
-public class JNIKiwixBool
+public class JNIKiwixString
 {
-  public boolean value;
+  public String value;
+
+  public JNIKiwixString(String value) {
+    this.value = value;
+  }
+
+  public JNIKiwixString() {
+    this("");
+  }
+
+  public String getValue() {
+    return value;
+  }
 }
