+++ app/src/main/java/org/kiwix/java_libkiwix/Manager.java
@@ -17,9 +17,9 @@
  * MA 02110-1301, USA.
  */
 
-package org.kiwix.kiwixlib;
+package org.kiwix.java_libkiwix;
 
-import org.kiwix.kiwixlib.Library;
+import org.kiwix.java_libkiwix.Library;
 
 public class Manager
 {
